git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/trunk/owsi-core@2783 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core-components/owsi-core-component-wicket-more/src/main/java/fr/openwide/core/wicket/more/markup/repeater/data/OddEvenDataView.java
@@ -12,8 +12,7 @@ public abstract class OddEvenDataView<T> extends DataView<T> {
 
 	private static final long serialVersionUID = 7579892102655985018L;
 
-	public OddEvenDataView(String id, final IDataProvider<T> dataProvider,
-			int itemsPerPage) {
+	public OddEvenDataView(String id, final IDataProvider<T> dataProvider, long itemsPerPage) {
 		super(id, dataProvider, itemsPerPage);
 	}
